+++ src/jump.jl
@@ -5,16 +5,9 @@ struct PhaseJump <: PhaseComponent
     jump_mask::BitMatrix
 end
 
-function phase(pjmp::PhaseJump, ctoa::CorrectedTOA, params::NamedTuple)::GQ
-    @assert length(params.JUMP) == size(pjmp.jump_mask)[1]
-    @assert ctoa.toa.index <= size(pjmp.jump_mask)[2]
-    if ctoa.toa.tzr
-        return dimensionless(0.0)
-    else
-        return dot(params.JUMP, @view(pjmp.jump_mask[:, ctoa.toa.index])) *
-               (params.F_ + params.F[1])
-    end
-end
+phase(pjmp::PhaseJump, ctoa::CorrectedTOA, params::NamedTuple)::GQ =
+    ctoa.toa.tzr ? dimensionless(0.0) :
+    (basis_dot(pjmp.jump_mask, params.JUMP, ctoa.toa.index) * (params.F_ + params.F[1]))
 
 function show(io::IO, jmp::PhaseJump)
     num_jumps = size(jmp.jump_mask)[1]