+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Move chi2 functions into a separate file `chi2.jl`
 - Move the higher order functions in `pyinter.jl` to `chi2.jl` and `likelihood.jl`
 - Updated `README` to use `pint2vela`
+- `basis_dot` function
 ## Fixed
 ## Removed